plugins: rename resources to http
Why:

* comply with wazo-notebook/file-structure.md
