+++ wazo_calld/plugins/applications/http.py
