+++ wazo_calld/plugins/conferences/plugin.py
@@ -5,7 +5,7 @@ from wazo_auth_client import Client as AuthClient
 from wazo_confd_client import Client as ConfdClient
 from wazo_amid_client import Client as AmidClient
 
-from .resources import (
+from .http import (
     ConferenceRecordResource,
     ParticipantMuteResource,
     ParticipantResource,
