+++ wazo_calld/plugins/status/plugin.py
@@ -1,7 +1,7 @@
 # Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from .resources import StatusResource
+from .http import StatusResource
 
 
 class Plugin:
