+++ wazo_calld/plugins/voicemails/plugin.py
@@ -7,7 +7,7 @@ from wazo_auth_client import Client as AuthClient
 from wazo_confd_client import Client as ConfdClient
 
 from .bus_consume import VoicemailsBusEventHandler
-from .resources import (
+from .http import (
     UserVoicemailFolderResource,
     UserVoicemailGreetingCopyResource,
     UserVoicemailGreetingResource,
